feat(chat): purpose-built ChatFamiliarView — retire the InspectorPane transplant (cave-yqrx)
The chat surface's Familiar tab mounted the two-mode InspectorPane shell
(tab="familiar") — a holdover from the retired inspector sidepanel whose
other mode is the memory rail. Extract the identity hero + capability
panel + their building blocks into src/components/chat-familiar-view.tsx,
a first-class surface that owns its own <section> landmark, scroll region,
and no-familiar empty state. The visual design from cave-7e1l/cave-aovo/
cave-w3us carries over unchanged.

- inspector-pane.tsx returns to memory-only: the Tab union and the
  tab/daemonRunning/onStartChat props are gone; the pane is the rail's
  memory section again (683 lines out).
- Purpose-built presence: offline familiars with a last_seen record now
  read "offline · last seen 2h ago" via the canonical relativeTime in a
  semantic <time>; the Runtime "scanned" row uses relativeTime too.
- Test pins ported: familiar-tab-hero/-bridges/-sections read the new
  file; the landmark pin asserts both surfaces label themselves;
  inspector-pane-polish pins the pane as memory-only; chat-surface pins
  the ChatFamiliarView mount.
