+++ src/components/chat-surface.test.ts
@@ -224,8 +224,8 @@ assert.match(
 );
 assert.match(
   chatSurface,
-  /scope === "familiar" \? \([\s\S]*?<InspectorPane familiar=\{activeFamiliar\} tab="familiar" daemonRunning=\{daemonRunning\} onStartChat=\{startFamiliarHeroChat\} \/>/,
-  "the familiar scope renders the capability panel for the active familiar",
+  /scope === "familiar" \? \([\s\S]*?<ChatFamiliarView familiar=\{activeFamiliar\} daemonRunning=\{daemonRunning\} onStartChat=\{startFamiliarHeroChat\} \/>/,
+  "the familiar scope renders the purpose-built familiar view for the active familiar",
 );
 
 assert.match(
